+++ src/Config.hs
@@ -201,9 +201,11 @@ sourceFile Basketball = "basketball"
 sourceFile Opera = "opera"
 sourceFile Blood = "blood"
 sourceFile Chiquito = "chiquito"
+sourceFile Computer = "computer"
 
 mapSource :: Text -> Name
 mapSource "chiquito" = 'Chiquito
+mapSource "computer" = 'Computer
 mapSource "horse" = 'Horse
 mapSource "address" = 'Address
 mapSource "basketball" = 'Basketball